requirements.txt
Pillow
bs4
langdetect
